-Updated to Terraria 1.2.3. -Version tick 2.7.

+++ Hooks/ChestModifySlotEventArgs.cs
@@ -5,12 +5,12 @@ using TShockAPI;
 
 namespace Terraria.Plugins.Common.Hooks {
   public class ChestModifySlotEventArgs: NetHookEventArgs {
-    public short ChestIndex { get; private set; }
-    public byte SlotIndex { get; private set; }
+    public int ChestIndex { get; private set; }
+    public int SlotIndex { get; private set; }
     public ItemData NewItem { get; private set; }
 
 
-    public ChestModifySlotEventArgs(TSPlayer player, short chestIndex, byte slotIndex, ItemData newItem): base(player) {
+    public ChestModifySlotEventArgs(TSPlayer player, int chestIndex, int slotIndex, ItemData newItem): base(player) {
       this.ChestIndex = chestIndex;
       this.SlotIndex = slotIndex;
       this.NewItem = newItem;
