+++ Hooks/ItemOwnerEventArgs.cs
@@ -5,11 +5,11 @@ using TShockAPI;
 
 namespace Terraria.Plugins.Common.Hooks {
   public class ItemOwnerEventArgs: NetHookEventArgs {
-    public short ItemIndex { get; private set; }
+    public int ItemIndex { get; private set; }
     public TSPlayer NewOwner { get; private set; }
 
 
-    public ItemOwnerEventArgs(TSPlayer player, short itemIndex, TSPlayer newOwner): base(player) {
+    public ItemOwnerEventArgs(TSPlayer player, int itemIndex, TSPlayer newOwner): base(player) {
       this.ItemIndex = itemIndex;
       this.NewOwner = newOwner;
     }
