+++ Hooks/TileEditEventArgs.cs
@@ -7,11 +7,11 @@ namespace Terraria.Plugins.Common.Hooks {
   public class TileEditEventArgs: TileLocationEventArgs {
     public TileEditType EditType { get; private set; }
     public BlockType BlockType { get; private set; }
-    public byte ObjectStyle { get; private set; }
+    public int ObjectStyle { get; private set; }
 
 
     public TileEditEventArgs(
-      TSPlayer player, TileEditType editType, DPoint location, BlockType blockType, byte objectStyle
+      TSPlayer player, TileEditType editType, DPoint location, BlockType blockType, int objectStyle
     ): base(player, location) {
       this.EditType = editType;
       this.BlockType = blockType;
